feat: refactor and add aliases
