+++ src/components/tripDetails/_CostItemTripDetails.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import { Graphic, Item, Meta, PrimaryText, SecondaryText, Text } from '@smui/list';
-	import type { Cost } from '../../interfaces/Trips';
+	import type { Cost } from '$tripDomain';
 	import { createEventDispatcher } from 'svelte';
 
 	export let cost: Cost;
