+++ src/domain/trip/ChartData.ts
@@ -1,0 +1,7 @@
+export interface ChartData {
+	labels: Array<string>;
+	datasets: Array<{
+		data: Array<number>;
+		backgroundColor: Array<string>;
+	}>;
+}
