+++ src/domain/trip/CostCategory.ts
@@ -1,0 +1,6 @@
+export interface CostCategory {
+	name: string;
+	totalAmount: number;
+	color: string;
+	icon: string;
+}
