+++ src/domain/trip/CostList.ts
@@ -1,0 +1,5 @@
+import type { Cost } from './Cost';
+
+export interface CostList extends Cost {
+	listDisabled?: boolean;
+}
