+++ src/domain/trip/index.ts
@@ -1,0 +1,5 @@
+export type { ChartData } from './ChartData';
+export type { Cost } from './Cost';
+export type { CostCategory } from './CostCategory';
+export type { CostList } from './CostList';
+export type { TravelData } from './TravelData';
