+++ src/domain/user/User.ts
@@ -1,0 +1,3 @@
+export interface User {
+	name: string;
+}
