+++ src/stores/index.ts
@@ -1,0 +1,1 @@
+export { allTrips, currentTrip } from './tripsStore';
