+++ tsconfig.json
@@ -10,7 +10,11 @@
 		"sourceMap": true,
 		"strict": true,
 		"paths": {
-			"$components": ["./src/components"]
+			"$components": ["./src/components"],
+			"$stores": ["./src/stores"],
+			"$utils": ["./src/utils"],
+			"$tripDomain": ["./src/domain/trip"],
+			"$userDomain": ["./src/domain/user"]
 		}
 	}
 	// Path aliases are handled by https://kit.svelte.dev/docs/configuration#alias
