+++ vite.config.ts
@@ -6,7 +6,11 @@ export default defineConfig({
 	plugins: [sveltekit()],
 	resolve: {
 		alias: {
-			$components: path.resolve('./src/components')
+			$components: path.resolve('./src/components'),
+			$stores: path.resolve('./src/stores'),
+			$utils: path.resolve('./src/utils'),
+			$tripDomain: path.resolve('./src/domain/trip'),
+			$userDomain: path.resolve('./src/domain/user')
 		}
 	}
 });
